requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==5.1.0
carla==0.9.13
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clearml==1.4.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docker==5.0.3
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
furl==2.1.3
future==0.18.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.6.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
orderedmultidict==1.0.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkg_resources==0.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.1.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.3
testpath==0.6.0
torch-utils==0.1.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
widgetsnbextension==3.6.0
zipp==3.7.0
